moved utility classes into their own files
