+++ CaesarCipher.java
@@ -3,231 +3,8 @@ import java.util.*;
 import java.util.stream.*;
 import java.lang.Character;
 
-
-class CliInputs
-{
-    public static int key;
-    public static String action;
-    public static String text;
-    public static String filePath;
-
-
-    public CliInputs(String[] args) 
-    {
-        String action = args[0];
-        boolean setActionSuccess = setAction(action);
-        handleError(setActionSuccess, "Error with action");
-
-        int key = Integer.parseInt(args[1]);
-        boolean setKeySuccess = setKey(key);
-        handleError(setKeySuccess, "Error with encryption key");
-
-        if (action.equals("encrypt")
-        || action.equals("decrypt"))
-        {
-            String text = args[2];
-            boolean setTextSuccess = setText(text);
-            handleError(setTextSuccess, "Error with text");
-
-        }
-        else if (action.equals("encryptFile")
-        || action.equals("decryptFile"))
-        {
-            String filePath = args[2];
-            boolean setFilePathSuccess = setFilePath(filePath);
-            handleError(setFilePathSuccess, "Error with file path");
-        }
-    }
-
-
-    private void handleError(boolean status, String msg)
-    {
-        // figure out how to properly do this
-        if (!status)
-        {
-            // throw new Exception(msg);
-        }
-    }
-
-
-    private static boolean setKey(int k) 
-    {
-        try 
-        {
-            int keyValue;
-            if (k > 26)
-            {
-                keyValue = k - 26;
-            }
-            else
-            {
-                keyValue = k;
-            }
-            key = k;        
-        }
-        catch(Exception e) 
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-
-    private static boolean setAction(String a) 
-    {
-        // check to see if the action is valid
-        try 
-        {
-            action = a;
-        }
-        catch(Exception e) 
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-
-    private static boolean setText(String t)
-    {
-        try 
-        {
-            text = t;
-        }
-        catch(Exception e) 
-        {
-            return false;
-        }
-
-        return true;
-    }
-
-
-    private static boolean setFilePath(String p)
-    {
-        // check to see if file exists/valid file path
-         try 
-        {
-            filePath = p;
-        }
-        catch(Exception e) 
-        {
-            return false;
-        }
-
-        return true;
-    }
-}
-
-class FileHandler
-{
-    private static String filepath;
-
-    public FileHandler(String f)
-    {
-        boolean fileExists = setFile(f);
-        if (!fileExists)
-        {
-            // handle error here
-            return;
-        }
-    }
-
-
-    private static String createOutputFile()
-    {
-        String outputFileName = "";
-        String filenameNoExtension = filepath.replace(".txt" ,"");
-        
-        try 
-        {
-            outputFileName = filenameNoExtension + "_output.txt";
-            File fileObj = new File(outputFileName);
-            boolean fileCreated = fileObj.createNewFile();
-
-            if (!fileCreated) 
-            {
-                System.out.println("error creating output file");
-            }
-        } 
-        catch (IOException e) 
-        {
-            System.out.println("An error occurred.");
-            e.printStackTrace();
-        }
-
-        return outputFileName;
-    }
-
-
-    public static String writeOutputFile(String data)
-    {
-        String outputFile = createOutputFile();
-
-        try 
-        {
-            FileWriter file = new FileWriter(outputFile);
-            file.write(data);
-            file.close();
-
-        } 
-        catch (IOException e) 
-        {
-            System.out.println("An error occurred.");
-            e.printStackTrace();
-        }
-        return "file path goes here";
-    }
-
-
-    public static String readInputFile()
-    {
-        String data = "";
-
-        try
-        {
-            File fileObj = new File(filepath);
-            Scanner fileReader = new Scanner(fileObj);
-            
-            while (fileReader.hasNextLine()) 
-            {
-                data += fileReader.nextLine();
-            }
-
-            fileReader.close();
-        
-        }
-        catch (FileNotFoundException e)
-        {
-            System.out.println("An error occurred.");
-            e.printStackTrace();
-
-        }
-        return data;
-    }
-
-
-    private static boolean setFile(String filePath)
-    {
-        File f = new File(filePath);
-        boolean fileExists = f.exists();
-        boolean fileCanRead = f.canRead();
-
-        if (fileExists && fileCanRead)
-        {
-            filepath = filePath;
-            return true;
-        }
-        else
-        {
-            return false;
-        }
-    }
-}
-
-
+import Utilities.CliInputs;
+import Utilities.FileHandler;
 
 public class CaesarCipher 
 {
